feat(form): add palette color form component
