+++ components/form/index.ts
@@ -1,2 +1,3 @@
 export * from "./search-color"
 export * from "./palette-menu"
+export * from "./palette-color-form"
